Fix moving around groovidelics/funkotronics (Found with AddressSanitizer)

+++ src/DETHRACE/common/world.h
@@ -123,7 +123,7 @@ void ControlBoundFunkGroove(int pSlot_number, float pValue);
 
 float ControlBoundFunkGroovePlus(int pSlot_number, float pValue);
 
-void ShiftBoundGrooveFunks(char* pStart, char* pEnd, int pDelta);
+void ShiftBoundGrooveFunks(char* pStart, char* pEnd, size_t pDelta);
 
 tFunkotronic_spec* AddNewFunkotronic();
